# Conflicts:
#	src/main/java/com/catalyst/collector/entities/Collectible.java

+++ src/main/java/com/catalyst/collector/services/impl/CollectionsServiceImpl.java
@@ -8,20 +8,6 @@ import com.catalyst.collector.entities.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.catalyst.collector.daos.CollectionsDao;
-import com.catalyst.collector.entities.Color;
-import com.catalyst.collector.entities.Category;
-import com.catalyst.collector.entities.Keyword;
-import com.catalyst.collector.entities.Collectible;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import com.catalyst.collector.daos.CollectionsDao;
-import java.util.List;
-
-import com.catalyst.collector.entities.Age;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.catalyst.collector.daos.CollectionsDao;
 import com.catalyst.collector.services.CollectionsService;
 
 @Service
@@ -34,32 +20,26 @@ public class CollectionsServiceImpl implements CollectionsService {
 		this.collectionsDao = collectionsDao;
 	}
 
-
-
 	@Override
 	public ArrayList<Age> getAgeTypes(){
 		return collectionsDao.getAgeTypes();
 	}
 
 	@Override
-	public boolean addAge(Age age) {
+	public void addAge(Age age) {
 		String ageString = age.getAge();
-		if(ageString.length() <= 255 && !ageString.matches(".*\\d.*")) { //Maximum of 255 characters for an age, no digits allowed
+		if(ageString != null && ageString.length() <= 255 && !ageString.matches(".*\\d.*") && !ageString.trim().equals("")) { //Maximum of 255 characters for an age, no digits allowed
 			collectionsDao.addAge(age);
-			return true;
 		}
-		return false;
 	}
 
 
 	@Override
-	public boolean updateAge(Age age){
+	public void updateAge(Age age){
 		String ageString = age.getAge();
-		if(ageString.length() <= 255 && !ageString.matches(".*\\d.*")) { //Maximum of 255 characters for an age, no digits allowed
+		if(ageString != null && ageString.length() <= 255 && !ageString.matches(".*\\d.*") && !ageString.trim().equals("")) { //Maximum of 255 characters for an age, no digits allowed
 			collectionsDao.updateAge(age);
-			return true;
 		}
-		return false;
 
 	}
 
@@ -81,14 +61,18 @@ public class CollectionsServiceImpl implements CollectionsService {
 
 	@Override
 	public boolean addCategory(Category category) {
-		if(category.getName() == null || ((category.getName()).trim()).equals(""))
+		
+		if(category.getCategory() == null || ((category.getCategory()).trim()).equals("")|| !category.getCategory().matches("^[a-zA-Z0-9]*$"))
 		{
 			return false;
 		}
-		if((category.getName()).length() > 255)
+		if((category.getCategory()).length() > 255)
 		{
 			return false;
 		}
+		String cat = category.getCategory();
+		cat.toLowerCase();
+		category.setCategory(cat);
 		return collectionsDao.addCategory(category);
 
 	}
@@ -96,20 +80,31 @@ public class CollectionsServiceImpl implements CollectionsService {
 
 	@Override
 	public boolean updateCategory(int id, Category category) {
+
 		category.setId(id);
-		if(category.getName() == null || ((category.getName()).trim()).equals(""))
+       
+		if(category.getCategory() == null || ((category.getCategory()).trim()).equals("")  || !category.getCategory().matches("^[a-zA-Z0-9]*$") )
 		{
 			return false;
 		}
-		if((category.getName()).length() > 255)
+		if(category.getCategory().matches("\\s") 
+				|| (category.getCategory()).length() > 255 
+					|| category.getId()<1)
 		{
 			return false;
 		}
+		String cat = category.getCategory();
+		cat.toLowerCase();
+		category.setCategory(cat);
 		return collectionsDao.updateCategory(category);
 	}
 
 	@Override
 	public boolean deleteCategory(int id) {
+		if(id<1)
+		{
+			return false;
+		}
 		return collectionsDao.deleteCategory(id);
 	}
 
@@ -119,12 +114,10 @@ public class CollectionsServiceImpl implements CollectionsService {
 	}
 	@Override
 	public boolean addColor(Color addedColor) {
-		try{
-			collectionsDao.addColor(addedColor);
-		}
-		catch(Exception e){
-			return false;
-		}
+		 if (addedColor.getColor() == null || addedColor.getColor().trim().equals("") || addedColor.getColor().length() > 255){
+	            return false;
+		 }
+		collectionsDao.addColor(addedColor);
 		return true;
 		}
 	@Override
@@ -139,38 +132,23 @@ public class CollectionsServiceImpl implements CollectionsService {
 	}
 
 	@Override
-	public boolean updateColor(int id, String color){
-		try{
-		Color c = getColor(id);
-		c.setColor(color);
+	public boolean updateColor(Color c){
+		 if (c.getColor() == null || c.getColor().trim().equals("") || c.getColor().length() > 255){
+	            return false;
+		 }
 		collectionsDao.updateColor(c);
-		}
-		catch(Exception e){
-			return false;
-			}
 		return true;
 	}
 	public Color getColor(int id){
-		List<Color> colors = getColorList();
-		for(Color c: colors){
-			if(c.getId() == id){
-				return c;
-			}
-		}
-		return null;
+		return collectionsDao.getColor(id);
 
-	}
-
-	@Override
-	public Color getByColorId(int colorId){
-		return collectionsDao.getColor(colorId);
 	}
 
     @Override
     public ArrayList<Keyword> getAllKeywords() {
         return collectionsDao.getAllKeywords();
     }
-    
+
     @Override
     public ArrayList<Keyword> getKeywordsByLetter(String letter){
     	return collectionsDao.getKeywordsByLetter(letter);
@@ -178,7 +156,7 @@ public class CollectionsServiceImpl implements CollectionsService {
 
     @Override
     public boolean addKeyword(Keyword keyword) {
-        if (keyword.getWord().length() < 1 || keyword.getWord().length() > 255)
+        if (!keyword.isValid())
             return false;
 
         collectionsDao.addKeyword(keyword);
@@ -187,7 +165,7 @@ public class CollectionsServiceImpl implements CollectionsService {
 
     @Override
     public boolean updateKeyword(Keyword keyword) {
-        if (keyword.getWord() == null || keyword.getWord().equals("") || keyword.getWord().length() > 255)
+        if (!keyword.isValid())
             return false;
         collectionsDao.updateKeyword(keyword);
         return true;
@@ -195,12 +173,15 @@ public class CollectionsServiceImpl implements CollectionsService {
 
     @Override
     public void removeKeyword(Integer id) {
+        if (id < 0)
+            return;
         collectionsDao.removeKeyword(id);
     }
 
 	@Override
-	public void addCollectible(Collectible collectible) {
+	public boolean addCollectible(Collectible collectible) {
 		collectionsDao.addCollectible(collectible);
+		return true;
 	}
 
 	@Override
@@ -242,6 +223,11 @@ public class CollectionsServiceImpl implements CollectionsService {
 	@Override
 	public Collectible getCollectible(Integer id) {
 		return collectionsDao.getCollectible(id);
+	}
+
+	@Override
+	public boolean removeCollectible(int id) {
+		return collectionsDao.removeCollectible(id);
 	}
 
 }
